Addition of Bezkorupce.cz reporting site
Addition of Bezkorupce.cz, a site dealing with reporting and
whistleblowing on conflict of interests, public management property,
transparency of municipal governments in the Czech Republic.

It is managed by the NGO Oživení who promote the transparency in public
administration and sustainable development. It uses the Global Leaks
platform.

Including first-party "proof" url.